server/sentiment_service/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
transformers==4.31.0
torch==2.5.0
